@@ -3,7 +3,7 @@ package io.github.adrianulbona.kata.cc_01_arrays_strings;
 import java.util.Arrays;
 import java.util.function.BiPredicate;
 
-public class _02_StringPermutationChecker implements BiPredicate<char[], char[]> {
+class _02_StringPermutationChecker implements BiPredicate<char[], char[]> {
 
     @Override
     public boolean test(char[] s1, char[] s2) {
