@@ -2,7 +2,7 @@ package io.github.adrianulbona.kata.cc_02_lists;
 
 import java.util.function.BiFunction;
 
-public class _02_KLastFinder implements BiFunction<Integer, Node, Node> {
+class _02_KLastFinder implements BiFunction<Integer, Node, Node> {
 
     @Override
     public Node apply(Integer k, Node node) {
