@@ -3,7 +3,7 @@ package io.github.adrianulbona.kata.cc_02_lists;
 import java.util.function.BiFunction;
 import java.util.function.Predicate;
 
-public class _04_Partitioner implements BiFunction<Node, Integer, Node> {
+class _04_Partitioner implements BiFunction<Node, Integer, Node> {
 
     @Override
     public Node apply(Node node, Integer key) {
