@@ -1,12 +1,6 @@
 package io.github.adrianulbona.kata.cc_03_stacks_queues;
 
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-class _04_Queue<E> implements Queue<E> {
-
-    private final Stack<E> stack;
-    private final Stack<E> buffer;
+record _04_Queue<E>(Stack<E> stack, Stack<E> buffer) implements Queue<E> {
 
     static <E> _04_Queue<E> create() {
         return new _04_Queue<>(StackImpl.create(), StackImpl.create());
