@@ -2,7 +2,7 @@ package io.github.adrianulbona.kata.cc_05_bits;
 
 import java.util.function.Predicate;
 
-public class _05_Pow2Checker implements Predicate<Integer> {
+class _05_Pow2Checker implements Predicate<Integer> {
 
     @Override
     public boolean test(Integer n) {
