@@ -2,7 +2,7 @@ package io.github.adrianulbona.kata.cc_05_bits;
 
 import java.util.function.BiFunction;
 
-public class _06_FlipMatcher implements BiFunction<Integer, Integer, Integer> {
+class _06_FlipMatcher implements BiFunction<Integer, Integer, Integer> {
 
     @Override
     public Integer apply(Integer source, Integer target) {
