@@ -2,7 +2,7 @@ package io.github.adrianulbona.kata.cc_05_bits;
 
 import java.util.function.Function;
 
-public class _07_PairWiseSwapper implements Function<Integer, Integer> {
+class _07_PairWiseSwapper implements Function<Integer, Integer> {
 
     @Override
     public Integer apply(Integer input) {
