@@ -1,14 +1,12 @@
 package io.github.adrianulbona.kata.cc_10_sort_search;
 
-import lombok.Data;
-
 import java.util.Arrays;
 import java.util.Comparator;
 import java.util.List;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
-public class _02_AnagramsSorter implements Function<List<String>, List<String>> {
+class _02_AnagramsSorter implements Function<List<String>, List<String>> {
 
     @Override
     public List<String> apply(List<String> anagrams) {
@@ -25,9 +23,6 @@ public class _02_AnagramsSorter implements Function<List<String>, List<String>> 
         return new SortedString(new String(chars), string);
     }
 
-    @Data
-    private static class SortedString {
-        private final String sorted;
-        private final String raw;
+    record SortedString(String sorted, String raw){
     }
 }
